Remove mathematically unsupported subctypep methods
They are not correct in general. The particular case I hit was
(subtypep '(cons t (not float))
          '(or (cons integer (not float)) (cons (not integer))))
which is true even though both subtypes with the OR types are
false.

This makes the system noticeably worse, e.g. it can no longer
determine that (rational 10) is not a subtype of (rational 11),
which makes it nonconforming, but at least it doesn't confidently
give wrong answers.