Use a method combination instead of demanding call-next-method
First off call-next-method is accident prone. In the course of
this rewrite I found several places that were returning unknown
immediately rather than properly letting other methods have a go.

Secondly for programmers to define new kinds, they'd need to use
SURELY and stuff. That increases the accidents and also means
that stuff has to be exported.

Doing it this way also emphasizes that methods should stand on
their own. That one I rewrote in pairwise demonstrates the problem
of thinking in terms of next methods. Seriously I don't know what
I was thinking there, if the next method knows it's not a subtype
let it say that.