Define package with typep, subtypep functions
Not sure about what name to use. I'm going with "-extrinsic" for
now because it kind of matches Khazern even though we don't have
an intrinsic system, and because this way any existing users of
the ctype package can keep using it. Although they need to add
client arguments for that anyway...